requirements.txt
cytoolz
attrs
frozendict
numpy
pandas
feather-format
